Delete League Working
Needed to change update leagues route and db to contain leagues that the user has excluded so that we do not fetch them upon updating our users leagues. MAX_COUNT in this same route was changed to be the same as league_ids length so that after delete we do not add another random league to take its place.